restructure with ugrid module
